+++ integration/hsm/hsm_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+package hsm
 
 import (
 	"context"
@@ -35,11 +35,8 @@ import (
 	"github.com/gravitational/teleport/lib/backend"
 	"github.com/gravitational/teleport/lib/backend/etcdbk"
 	"github.com/gravitational/teleport/lib/backend/lite"
-	"github.com/gravitational/teleport/lib/cloud"
 	"github.com/gravitational/teleport/lib/modules"
-	"github.com/gravitational/teleport/lib/service"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/utils"
 )
 
@@ -57,224 +54,10 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
-type teleportService struct {
-	name           string
-	log            utils.Logger
-	config         *servicecfg.Config
-	process        *service.TeleportProcess
-	serviceChannel chan *service.TeleportProcess
-	errorChannel   chan error
-}
+func newHSMAuthConfig(t *testing.T, storageConfig *backend.Config, log utils.Logger) *servicecfg.Config {
+	config := newAuthConfig(t, log)
 
-func newTeleportService(t *testing.T, config *servicecfg.Config, name string) *teleportService {
-	s := &teleportService{
-		config:         config,
-		name:           name,
-		log:            config.Log,
-		serviceChannel: make(chan *service.TeleportProcess, 1),
-		errorChannel:   make(chan error, 1),
-	}
-	t.Cleanup(func() {
-		require.NoError(t, s.Close(), "error while closing %s during test cleanup", name)
-	})
-	return s
-}
-
-func (t *teleportService) Close() error {
-	if t.process == nil {
-		return nil
-	}
-	if err := t.process.Close(); err != nil {
-		return trace.Wrap(err)
-	}
-	return trace.Wrap(t.process.Wait())
-}
-
-func (t *teleportService) start(ctx context.Context) {
-	go func() {
-		t.errorChannel <- service.Run(ctx, *t.config, func(cfg *servicecfg.Config) (service.Process, error) {
-			t.log.Debugf("(Re)starting %s", t.name)
-			svc, err := service.NewTeleport(cfg)
-			if err == nil {
-				t.log.Debugf("started %s, writing to serviceChannel", t.name)
-				t.serviceChannel <- svc
-			}
-			return svc, trace.Wrap(err)
-		})
-	}()
-}
-
-func (t *teleportService) waitForStart(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to start", t.name)
-	t.start(ctx)
-	select {
-	case t.process = <-t.serviceChannel:
-	case err := <-t.errorChannel:
-		return trace.Wrap(err)
-	case <-ctx.Done():
-		return trace.Wrap(ctx.Err(), "timed out waiting for %s to start", t.name)
-	}
-	t.log.Debugf("read %s from serviceChannel", t.name)
-	return t.waitForReady(ctx)
-}
-
-func (t *teleportService) waitForReady(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to be ready", t.name)
-	if _, err := t.process.WaitForEvent(ctx, service.TeleportReadyEvent); err != nil {
-		return trace.Wrap(err, "timed out waiting for %s to be ready", t.name)
-	}
-	// also wait for AuthIdentityEvent so that we can read the admin credentials
-	// and create a test client
-	if t.process.GetAuthServer() != nil {
-		if _, err := t.process.WaitForEvent(ctx, service.AuthIdentityEvent); err != nil {
-			return trace.Wrap(err, "timed out waiting for %s auth identity event", t.name)
-		}
-		t.log.Debugf("%s is ready", t.name)
-	}
-	return nil
-}
-
-func (t *teleportService) waitForRestart(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to restart", t.name)
-	// get the new process
-	select {
-	case t.process = <-t.serviceChannel:
-	case err := <-t.errorChannel:
-		return trace.Wrap(err)
-	case <-ctx.Done():
-		return trace.Wrap(ctx.Err(), "timed out waiting for %s to restart", t.name)
-	}
-
-	// wait for the new process to be ready
-	err := t.waitForReady(ctx)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	t.log.Debugf("%s successfully restarted", t.name)
-	return nil
-}
-
-func (t *teleportService) waitForShutdown(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to shut down", t.name)
-	select {
-	case err := <-t.errorChannel:
-		t.process = nil
-		return trace.Wrap(err)
-	case <-ctx.Done():
-		return trace.Wrap(ctx.Err(), "timed out waiting for %s to shut down", t.name)
-	}
-}
-
-func (t *teleportService) waitForLocalAdditionalKeys(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to have local additional keys", t.name)
-	clusterName, err := t.process.GetAuthServer().GetClusterName()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	hostCAID := types.CertAuthID{DomainName: clusterName.GetClusterName(), Type: types.HostCA}
-	for {
-		select {
-		case <-ctx.Done():
-			return trace.Wrap(ctx.Err(), "timed out waiting for %s to have local additional keys", t.name)
-		case <-time.After(250 * time.Millisecond):
-		}
-		ca, err := t.process.GetAuthServer().GetCertAuthority(ctx, hostCAID, true)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		hasUsableKeys, err := t.process.GetAuthServer().GetKeyStore().HasUsableAdditionalKeys(ctx, ca)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if hasUsableKeys {
-			break
-		}
-	}
-	t.log.Debugf("%s has local additional keys", t.name)
-	return nil
-}
-
-func (t *teleportService) waitForPhaseChange(ctx context.Context) error {
-	t.log.Debugf("Waiting for %s to change phase", t.name)
-	if _, err := t.process.WaitForEvent(ctx, service.TeleportPhaseChangeEvent); err != nil {
-		return trace.Wrap(err, "timed out waiting for %s to change phase", t.name)
-	}
-	t.log.Debugf("%s changed phase", t.name)
-	return nil
-}
-
-func (t *teleportService) AuthAddr(testingT *testing.T) utils.NetAddr {
-	addr, err := t.process.AuthAddr()
-	require.NoError(testingT, err)
-
-	return *addr
-}
-
-type TeleportServices []*teleportService
-
-func (s TeleportServices) forEach(f func(t *teleportService) error) error {
-	for i := range s {
-		if err := f(s[i]); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-	return nil
-}
-
-func (s TeleportServices) waitForStart(ctx context.Context) error {
-	return s.forEach(func(t *teleportService) error { return t.waitForStart(ctx) })
-}
-
-func (s TeleportServices) waitForRestart(ctx context.Context) error {
-	return s.forEach(func(t *teleportService) error { return t.waitForRestart(ctx) })
-}
-
-func (s TeleportServices) waitForLocalAdditionalKeys(ctx context.Context) error {
-	return s.forEach(func(t *teleportService) error { return t.waitForLocalAdditionalKeys(ctx) })
-}
-
-func (s TeleportServices) waitForPhaseChange(ctx context.Context) error {
-	return s.forEach(func(t *teleportService) error { return t.waitForPhaseChange(ctx) })
-}
-
-func newHSMAuthConfig(ctx context.Context, t *testing.T, storageConfig *backend.Config, log utils.Logger) *servicecfg.Config {
-	hostName, err := os.Hostname()
-	require.NoError(t, err)
-
-	config := servicecfg.MakeDefaultConfig()
-	config.PollingPeriod = 1 * time.Second
-	config.SSH.Enabled = false
-	config.Proxy.Enabled = false
-	config.ClientTimeout = time.Second
-	config.ShutdownTimeout = time.Minute
-	config.DataDir = t.TempDir()
-	config.Auth.ListenAddr.Addr = net.JoinHostPort(hostName, "0")
-	config.Auth.PublicAddrs = []utils.NetAddr{
-		{
-			AddrNetwork: "tcp",
-			Addr:        hostName,
-		},
-	}
-	config.Auth.ClusterName, err = services.NewClusterNameWithRandomID(types.ClusterNameSpecV2{
-		ClusterName: "testcluster",
-	})
-	require.NoError(t, err)
-	config.SetAuthServerAddress(config.Auth.ListenAddr)
-	config.Auth.StaticTokens, err = types.NewStaticTokens(types.StaticTokensSpecV2{
-		StaticTokens: []types.ProvisionTokenV1{
-			{
-				Roles: []types.SystemRole{"Proxy", "Node"},
-				Token: "foo",
-			},
-		},
-	})
-	require.NoError(t, err)
-	config.Log = log
-	if storageConfig != nil {
-		config.Auth.StorageConfig = *storageConfig
-	}
-	config.CircuitBreakerConfig = breaker.NoopBreakerConfig()
-	config.InstanceMetadataClient = cloud.NewDisabledIMDSClient()
+	config.Auth.StorageConfig = *storageConfig
 
 	if gcpKeyring := os.Getenv("TEST_GCP_KMS_KEYRING"); gcpKeyring != "" {
 		config.Auth.KeyStore.GCPKMS.KeyRing = gcpKeyring
@@ -283,32 +66,6 @@ func newHSMAuthConfig(ctx context.Context, t *testing.T, storageConfig *backend.
 		config.Auth.KeyStore = keystore.SetupSoftHSMTest(t)
 	}
 
-	return config
-}
-
-func newProxyConfig(ctx context.Context, t *testing.T, authAddr utils.NetAddr, log utils.Logger) *servicecfg.Config {
-	hostName, err := os.Hostname()
-	require.NoError(t, err)
-
-	config := servicecfg.MakeDefaultConfig()
-	config.PollingPeriod = 1 * time.Second
-	config.SetToken("foo")
-	config.SSH.Enabled = false
-	config.Auth.Enabled = false
-	config.Proxy.Enabled = true
-	config.Proxy.DisableWebInterface = true
-	config.Proxy.DisableWebService = true
-	config.Proxy.DisableReverseTunnel = true
-	config.Proxy.SSHAddr.Addr = net.JoinHostPort(hostName, "0")
-	config.Proxy.WebAddr.Addr = net.JoinHostPort(hostName, "0")
-	config.CachePolicy.Enabled = true
-	config.PollingPeriod = 1 * time.Second
-	config.ShutdownTimeout = time.Minute
-	config.DataDir = t.TempDir()
-	config.SetAuthServerAddress(authAddr)
-	config.CircuitBreakerConfig = breaker.NoopBreakerConfig()
-	config.InstanceMetadataClient = cloud.NewDisabledIMDSClient()
-	config.Log = log
 	return config
 }
 
@@ -374,21 +131,21 @@ func TestHSMRotation(t *testing.T) {
 	log := utils.NewLoggerForTests()
 
 	log.Debug("TestHSMRotation: starting auth server")
-	authConfig := newHSMAuthConfig(ctx, t, liteBackendConfig(t), log)
+	authConfig := newHSMAuthConfig(t, liteBackendConfig(t), log)
 	auth1 := newTeleportService(t, authConfig, "auth1")
 	t.Cleanup(func() {
 		require.NoError(t, auth1.process.GetAuthServer().GetKeyStore().DeleteUnusedKeys(ctx, nil))
 	})
-	teleportServices := TeleportServices{auth1}
+	allServices := teleportServices{auth1}
 
 	log.Debug("TestHSMRotation: waiting for auth server to start")
-	require.NoError(t, auth1.waitForStart(ctx))
+	require.NoError(t, auth1.start(ctx))
 
 	// start a proxy to make sure it can get creds at each stage of rotation
 	log.Debug("TestHSMRotation: starting proxy")
-	proxy := newTeleportService(t, newProxyConfig(ctx, t, auth1.AuthAddr(t), log), "proxy")
-	require.NoError(t, proxy.waitForStart(ctx))
-	teleportServices = append(teleportServices, proxy)
+	proxy := newTeleportService(t, newProxyConfig(t, auth1.authAddr(t), log), "proxy")
+	require.NoError(t, proxy.start(ctx))
+	allServices = append(allServices, proxy)
 
 	log.Debug("TestHSMRotation: sending rotation request init")
 	err := auth1.process.GetAuthServer().RotateCertAuthority(ctx, auth.RotateRequest{
@@ -397,7 +154,7 @@ func TestHSMRotation(t *testing.T) {
 		Mode:        types.RotationModeManual,
 	})
 	require.NoError(t, err)
-	require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+	require.NoError(t, allServices.waitForPhaseChange(ctx))
 
 	log.Debug("TestHSMRotation: sending rotation request update_clients")
 	err = auth1.process.GetAuthServer().RotateCertAuthority(ctx, auth.RotateRequest{
@@ -406,7 +163,7 @@ func TestHSMRotation(t *testing.T) {
 		Mode:        types.RotationModeManual,
 	})
 	require.NoError(t, err)
-	require.NoError(t, teleportServices.waitForRestart(ctx))
+	require.NoError(t, allServices.waitForRestart(ctx))
 
 	log.Debug("TestHSMRotation: sending rotation request update_servers")
 	err = auth1.process.GetAuthServer().RotateCertAuthority(ctx, auth.RotateRequest{
@@ -415,7 +172,7 @@ func TestHSMRotation(t *testing.T) {
 		Mode:        types.RotationModeManual,
 	})
 	require.NoError(t, err)
-	require.NoError(t, teleportServices.waitForRestart(ctx))
+	require.NoError(t, allServices.waitForRestart(ctx))
 
 	log.Debug("TestHSMRotation: sending rotation request standby")
 	err = auth1.process.GetAuthServer().RotateCertAuthority(ctx, auth.RotateRequest{
@@ -424,7 +181,7 @@ func TestHSMRotation(t *testing.T) {
 		Mode:        types.RotationModeManual,
 	})
 	require.NoError(t, err)
-	require.NoError(t, teleportServices.waitForRestart(ctx))
+	require.NoError(t, allServices.waitForRestart(ctx))
 }
 
 // Tests multiple CA rotations and rollbacks with 2 HSM auth servers in an HA configuration
@@ -444,15 +201,15 @@ func TestHSMDualAuthRotation(t *testing.T) {
 
 	// start a cluster with 1 auth server and a proxy
 	log.Debug("TestHSMDualAuthRotation: Starting auth server 1")
-	auth1Config := newHSMAuthConfig(ctx, t, storageConfig, log)
+	auth1Config := newHSMAuthConfig(t, storageConfig, log)
 	auth1 := newTeleportService(t, auth1Config, "auth1")
 	t.Cleanup(func() {
 		require.NoError(t, auth1.process.GetAuthServer().GetKeyStore().DeleteUnusedKeys(ctx, nil),
 			"failed to delete hsm keys during test cleanup")
 	})
-	authServices := TeleportServices{auth1}
-	teleportServices := append(TeleportServices{}, authServices...)
-	require.NoError(t, authServices.waitForStart(ctx), "auth service failed initial startup")
+	authServices := teleportServices{auth1}
+	allServices := append(teleportServices{}, authServices...)
+	require.NoError(t, authServices.start(ctx), "auth service failed initial startup")
 
 	log.Debug("TestHSMDualAuthRotation: Starting load balancer")
 	hostName, err := os.Hostname()
@@ -460,7 +217,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 	lb, err := utils.NewLoadBalancer(
 		ctx,
 		*utils.MustParseAddr(net.JoinHostPort(hostName, "0")),
-		auth1.AuthAddr(t),
+		auth1.authAddr(t),
 	)
 	require.NoError(t, err)
 	require.NoError(t, lb.Listen())
@@ -469,21 +226,21 @@ func TestHSMDualAuthRotation(t *testing.T) {
 
 	// start a proxy to make sure it can get creds at each stage of rotation
 	log.Debug("TestHSMDualAuthRotation: Starting proxy")
-	proxyConfig := newProxyConfig(ctx, t, utils.FromAddr(lb.Addr()), log)
+	proxyConfig := newProxyConfig(t, utils.FromAddr(lb.Addr()), log)
 	proxy := newTeleportService(t, proxyConfig, "proxy")
-	require.NoError(t, proxy.waitForStart(ctx), "proxy failed initial startup")
-	teleportServices = append(teleportServices, proxy)
+	require.NoError(t, proxy.start(ctx), "proxy failed initial startup")
+	allServices = append(allServices, proxy)
 
 	// add a new auth server
 	log.Debug("TestHSMDualAuthRotation: Starting auth server 2")
-	auth2Config := newHSMAuthConfig(ctx, t, storageConfig, log)
+	auth2Config := newHSMAuthConfig(t, storageConfig, log)
 	auth2 := newTeleportService(t, auth2Config, "auth2")
-	require.NoError(t, auth2.waitForStart(ctx))
+	require.NoError(t, auth2.start(ctx))
 	t.Cleanup(func() {
 		require.NoError(t, auth2.process.GetAuthServer().GetKeyStore().DeleteUnusedKeys(ctx, nil))
 	})
 	authServices = append(authServices, auth2)
-	teleportServices = append(teleportServices, auth2)
+	allServices = append(allServices, auth2)
 
 	// make sure the admin identity used by tctl works
 	getAdminClient := func() *auth.Client {
@@ -493,7 +250,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		require.NoError(t, err)
 		tlsConfig, err := identity.TLSConfig(nil)
 		require.NoError(t, err)
-		authAddrs := []utils.NetAddr{auth2.AuthAddr(t)}
+		authAddrs := []utils.NetAddr{auth2.authAddr(t)}
 		clt, err := auth.NewClient(client.Config{
 			Addrs: utils.NetAddrsToStrings(authAddrs),
 			Credentials: []client.Credentials{
@@ -518,7 +275,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseInit,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+				require.NoError(t, allServices.waitForPhaseChange(ctx))
 				require.NoError(t, authServices.waitForLocalAdditionalKeys(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
@@ -527,7 +284,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateClients,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -535,7 +292,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateServers,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -543,7 +300,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseStandby,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -562,7 +319,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 	}
 
 	// Safe to send traffic to new auth server now that a full rotation has been completed.
-	lb.AddBackend(auth2.AuthAddr(t))
+	lb.AddBackend(auth2.authAddr(t))
 
 	// load balanced client shoud work with either backend
 	getAdminClient = func() *auth.Client {
@@ -610,7 +367,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseInit,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+				require.NoError(t, allServices.waitForPhaseChange(ctx))
 				require.NoError(t, authServices.waitForLocalAdditionalKeys(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
@@ -619,7 +376,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseRollback,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -627,7 +384,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseStandby,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -635,7 +392,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseInit,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+				require.NoError(t, allServices.waitForPhaseChange(ctx))
 				require.NoError(t, authServices.waitForLocalAdditionalKeys(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
@@ -644,7 +401,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateClients,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -652,7 +409,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseRollback,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -660,7 +417,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseStandby,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -668,7 +425,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseInit,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+				require.NoError(t, allServices.waitForPhaseChange(ctx))
 				require.NoError(t, authServices.waitForLocalAdditionalKeys(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
@@ -677,7 +434,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateClients,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -685,7 +442,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateServers,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -693,7 +450,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseRollback,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -701,7 +458,7 @@ func TestHSMDualAuthRotation(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseStandby,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -731,14 +488,14 @@ func TestHSMMigrate(t *testing.T) {
 
 	// start a dual auth non-hsm cluster
 	log.Debug("TestHSMMigrate: Starting auth server 1")
-	auth1Config := newHSMAuthConfig(ctx, t, storageConfig, log)
+	auth1Config := newHSMAuthConfig(t, storageConfig, log)
 	auth1Config.Auth.KeyStore = keystore.Config{}
 	auth1 := newTeleportService(t, auth1Config, "auth1")
-	auth2Config := newHSMAuthConfig(ctx, t, storageConfig, log)
+	auth2Config := newHSMAuthConfig(t, storageConfig, log)
 	auth2Config.Auth.KeyStore = keystore.Config{}
 	auth2 := newTeleportService(t, auth2Config, "auth2")
-	require.NoError(t, auth1.waitForStart(ctx))
-	require.NoError(t, auth2.waitForStart(ctx))
+	require.NoError(t, auth1.start(ctx))
+	require.NoError(t, auth2.start(ctx))
 
 	log.Debug("TestHSMMigrate: Starting load balancer")
 	hostName, err := os.Hostname()
@@ -746,8 +503,8 @@ func TestHSMMigrate(t *testing.T) {
 	lb, err := utils.NewLoadBalancer(
 		ctx,
 		*utils.MustParseAddr(net.JoinHostPort(hostName, "0")),
-		auth1.AuthAddr(t),
-		auth2.AuthAddr(t),
+		auth1.authAddr(t),
+		auth2.authAddr(t),
 	)
 	require.NoError(t, err)
 	require.NoError(t, lb.Listen())
@@ -756,9 +513,9 @@ func TestHSMMigrate(t *testing.T) {
 
 	// start a proxy to make sure it can get creds at each stage of migration
 	log.Debug("TestHSMMigrate: Starting proxy")
-	proxyConfig := newProxyConfig(ctx, t, utils.FromAddr(lb.Addr()), log)
+	proxyConfig := newProxyConfig(t, utils.FromAddr(lb.Addr()), log)
 	proxy := newTeleportService(t, proxyConfig, "proxy")
-	require.NoError(t, proxy.waitForStart(ctx))
+	require.NoError(t, proxy.start(ctx))
 
 	// make sure the admin identity used by tctl works
 	getAdminClient := func() *auth.Client {
@@ -768,7 +525,7 @@ func TestHSMMigrate(t *testing.T) {
 		require.NoError(t, err)
 		tlsConfig, err := identity.TLSConfig(nil)
 		require.NoError(t, err)
-		authAddrs := []utils.NetAddr{auth2.AuthAddr(t)}
+		authAddrs := []utils.NetAddr{auth2.authAddr(t)}
 		clt, err := auth.NewClient(client.Config{
 			Addrs: utils.NetAddrsToStrings(authAddrs),
 			Credentials: []client.Credentials{
@@ -788,18 +545,18 @@ func TestHSMMigrate(t *testing.T) {
 	require.NoError(t, testClient(clt))
 
 	// Phase 1: migrate auth1 to HSM
-	lb.RemoveBackend(auth1.AuthAddr(t))
+	lb.RemoveBackend(auth1.authAddr(t))
 	auth1.process.Close()
 	require.NoError(t, auth1.waitForShutdown(ctx))
 	auth1Config.Auth.KeyStore = keystore.SetupSoftHSMTest(t)
 	auth1 = newTeleportService(t, auth1Config, "auth1")
-	require.NoError(t, auth1.waitForStart(ctx))
+	require.NoError(t, auth1.start(ctx))
 
 	clt = getAdminClient()
 	require.NoError(t, testClient(clt))
 
-	authServices := TeleportServices{auth1, auth2}
-	teleportServices := TeleportServices{auth1, auth2, proxy}
+	authServices := teleportServices{auth1, auth2}
+	allServices := teleportServices{auth1, auth2, proxy}
 
 	stages := []struct {
 		targetPhase string
@@ -808,7 +565,7 @@ func TestHSMMigrate(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseInit,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForPhaseChange(ctx))
+				require.NoError(t, allServices.waitForPhaseChange(ctx))
 				require.NoError(t, authServices.waitForLocalAdditionalKeys(ctx))
 				clt := getAdminClient()
 				require.NoError(t, testClient(clt))
@@ -817,7 +574,7 @@ func TestHSMMigrate(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateClients,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -825,7 +582,7 @@ func TestHSMMigrate(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseUpdateServers,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -833,7 +590,7 @@ func TestHSMMigrate(t *testing.T) {
 		{
 			targetPhase: types.RotationPhaseStandby,
 			verify: func(t *testing.T) {
-				require.NoError(t, teleportServices.waitForRestart(ctx))
+				require.NoError(t, allServices.waitForRestart(ctx))
 				clt = getAdminClient()
 				require.NoError(t, testClient(clt))
 			},
@@ -852,18 +609,18 @@ func TestHSMMigrate(t *testing.T) {
 	}
 
 	// Safe to send traffic to new auth1 again
-	lb.AddBackend(auth1.AuthAddr(t))
+	lb.AddBackend(auth1.authAddr(t))
 
 	// Phase 2: migrate auth2 to HSM
-	lb.RemoveBackend(auth2.AuthAddr(t))
+	lb.RemoveBackend(auth2.authAddr(t))
 	auth2.process.Close()
 	require.NoError(t, auth2.waitForShutdown(ctx))
 	auth2Config.Auth.KeyStore = keystore.SetupSoftHSMTest(t)
 	auth2 = newTeleportService(t, auth2Config, "auth2")
-	require.NoError(t, auth2.waitForStart(ctx))
+	require.NoError(t, auth2.start(ctx))
 
-	authServices = TeleportServices{auth1, auth2}
-	teleportServices = TeleportServices{auth1, auth2, proxy}
+	authServices = teleportServices{auth1, auth2}
+	allServices = teleportServices{auth1, auth2, proxy}
 
 	clt = getAdminClient()
 	require.NoError(t, testClient(clt))
@@ -880,6 +637,6 @@ func TestHSMMigrate(t *testing.T) {
 	}
 
 	// Safe to send traffic to new auth2 again
-	lb.AddBackend(auth2.AuthAddr(t))
+	lb.AddBackend(auth2.authAddr(t))
 	require.NoError(t, testClient(clt))
 }
